Delete LICENSE
I have made up my mind and decided that it's too big of a hassle for me to license everything properly, because I would like to license the text as well, but because code and text are contained in the same file I've decided to just not care at all. I'm sorry but I can't do this.

But you're allowed to look at my code and learn from it :3
I can't disallow that, can I? That's why it's open source.